Removed Flight and Flight dependent behaviors from my branch. Love the framework, bros., but can't stand the magic of bindings everywhere. I'll listen to arguments for putting it back in, but for dev at this point I'm taking it out. I just think it allows us to take shortcuts that we shouldn't be taking for now.
Changed the way measurement works. If you set the public width/height props, it sets the explicitWidth/height props and invalidates. Then in the measurement phase, if explicitWidth/height are set, it sets the _width and _height props to those. If they aren't set, it goes ahead and uses the layout's measure algorithm, which sets the _width and _height props. This way the _width and _height are set in measure, before the layout phase.